refactor: update namespaces, require and use statements
