+++ src/upload/catalog/controller/extension/payment/webpay_rest.php
@@ -1,7 +1,7 @@
 <?php
 require_once DIR_SYSTEM . '/library/Transbank/vendor/autoload.php';
-use Transbank\Opencart\Webpay\TransbankSdkWebpay;
-use Transbank\Opencart\Webpay\LogHandler;
+use Transbank\Opencart\Webpay\Utils\TransbankSdkWebpay;
+use Transbank\Opencart\Webpay\Utils\LogHandler;
 use Transbank\Webpay\WebpayPlus\Responses\TransactionCommitResponse;
 
 class ControllerExtensionPaymentWebpayRest extends Controller {
